20231127commit

+++ delete_comment.php
@@ -14,14 +14,31 @@ if (isset($_GET['commentId'])) {
 $commentFile = "comments_$postId.txt";
 
 if (file_exists($commentFile)) {
-    $comments = file($commentFile);
-    unset($comments[$commentId]);
-    file_put_contents($commentFile, implode("", $comments));
-    if (empty($comments)) {
-      unlink($commentFile);
+  $comments = file($commentFile);
+  $found = false;
+  $commentIndex = 0;
+
+  while ($commentIndex < count($comments)) {
+    $comment = $comments[$commentIndex];
+    list($currentCommentId, $comment) = explode(':', $comment, 2);
+
+    if ($currentCommentId == $commentId) {
+      $found = true;
+      unset($comments[$commentIndex]);
+      file_put_contents($commentFile, implode("", $comments));
+
+      if (empty($comments)) {
+        unlink($commentFile);
+      }
+
+      header("Location: post_detail.php?postId=$postId");
+      exit();
+    }
+
+    $commentIndex++;
+  }
+
+  if (!$found) {
+    die('コメントが見つかりません。');
   }
 }
-
-header("Location: post_detail.php?id=$postId");
-exit();
-?>
