+++ delete_post.php
@@ -6,18 +6,34 @@ if (isset($_GET['postId'])) {
 }
 
 $postFile = 'posts.txt';
+
 if (file_exists($postFile)) {
-    $posts = file($postFile);
+  $posts = file($postFile);
+  $found = false;
+  $postIndex = 0;
 
-    unset($posts[$postId]);
-    file_put_contents($postFile, implode("", $posts));
+  while ($postIndex < count($posts)) {
+    $post = $posts[$postIndex];
+    list($currentPostId, $title, $message) = explode(':', $post, 3);
 
-    $commentFile = "comments_$postId.txt";
-    if (file_exists($commentFile)) {
+    if ($currentPostId == $postId) {
+      $found = true;
+      unset($posts[$postIndex]);
+      file_put_contents($postFile, implode("", $posts));
+
+      $commentFile = "comments_$postId.txt";
+      if (file_exists($commentFile)) {
         unlink($commentFile);
-    }
-}
+      }
 
-header("Location: index.php");
-exit();
-?>
+      header("Location: index.php");
+      exit();
+    }
+
+    $postIndex++;
+  }
+
+  if (!$found) {
+    die('投稿が見つかりません。');
+  }
+}
